MerchantGuideToTheGalaxy/src/main/java/com/amaap/merchantguide/domain/service/io/reader/IntergalacticTransactionDataReader.java
package com.amaap.merchantguide.domain.service.io.reader;

import java.io.BufferedReader;
import java.io.File;
import java.io.FileReader;
import java.io.IOException;

public class IntergalacticTransactionDataReader {
    public boolean readTransactionData(File file) {
        try (BufferedReader reader = new BufferedReader(new FileReader(file))) {
            String line;
            while ((line = reader.readLine()) != null) {
                System.out.println(line);
            }
        } catch (IOException e) {
            System.err.println("Error: " + e.getMessage());
            return false;
        }
        return true;
    }
}
